+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## NEXT
+
+### Breaking changes
+
+ * port / host / tags / namespace can no longer be set on the instance to allow thread-safety [#87][] by [@grosser][] 
+
 ## 3.3.0 / 2018.02.04
 
  * [FEATURE] Add distribution support (beta). See [#72][].